classification test needs to be run last due to problems with tf sessions and multiprocess
